requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
aiohttp==3.8.1
    # via py-cord
aiosignal==1.2.0
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==22.1.0
    # via aiohttp
charset-normalizer==2.1.1
    # via aiohttp
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via yarl
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
py-cord==2.1.3
    # via -r requirements.in
python-dotenv==0.21.0
    # via -r requirements.in
yarl==1.8.1
    # via aiohttp
